Adding Basic JSON Parsing
The system is not capable of reading in a simple JSON specification file
which allows configuration of some global parameters (screen size,
title, input device mapping) and state machine parameters (timeout and
command conditional advancement which will map to the input device
mapping - and visual and audio stimuli).

Major additions/test that still need to be made/done:
-Does the WWW call block? If not, what happens if it doesn't complete in
time? Try a very large file to test this.
-Error checking for basic JSON parsing (bracket problems, etc) should
end program as well as existing error checks
-The camera space is not currently equal to the game coordinate space.
These should be made equal by moving the camera (and possibly
rotating/flipping it)
-Test the scaling parameters on the images
-Test that the audio will play when called
-Add the input adapters which are capable of calling up current commands
which are active given a set of inputs - these commands are then used to
drive state machine conditions...
-Build the basic state machine structure and test some transitions
-Build the actual task to make sure no important features have been left
out (cross event conditionals, for instance)