Added Notification Manager and Other
-Added notification manager for prompting small messages in bottom left
corner
-Changed global pause to allow notifications by not modifying culling
mask
-Added prefabs for primary scene objects
-Changes input commands to accept only arrays of command strings (they
can have just one value) which are "or"ed together to determine the
result
-Added cumulative input command which does not reset its timer when the
button is released. This means if you press the button over and over it
will build up time until the threshold is reached.
-Added DisableGlobalPause stimulus option which will, when that state is
active, disable the ability to pause.
-Reordered state machine so end state evaluation is done before commands
are processed. This was in an attempt to debug some problems as the
order is arbitrary.
-Added public visibility GlobalPauseEnabled property s.t. other objects
can change pause state (for pause disable stimuli)