Expression & Massive Changes - See Description
-Added NCalc for evaluating boolean expressions
-Added Variable Engine for storing cumulative button press condition
variables (this is the only condition currently compatible with
variables, but others could be added later)
-Made Notification Manager static so references aren't needed when
calling it.
-Added notification text optional input to all conditions s.t. they can
pop up a notification when they occur if desired
-Added ExpressionCondition which can be used to evaluate expressions
using VariableEngine variables.
-Made GlobalPauseDisableStimuli no longer inherit from monobehaviour as
it wasn't necessary (the class can just be called)
-Reordered system state machine (again) s.t. commands are processed
first. This is a reversal of a debug version where they were processed
second and allows the program to end gracefully without null reference
exceptions.
-Made system state machine execute all possible states before finishing
updating. This means that if a state is simply "passed through" it won't
show its stimulus. This leads to a vulnerability where you can create
infinite state loops where their conditions are true by default and the
program locks up, but this is expected when you make an infinite loop in
the specification. This is the only known source of program locking so
if it's seen, that's probably the cause.
-Made pause set culling mask to only show UI (no stimuli) then revert
back to everything when unpaused
-Updated task configuration to have the appropriate procedure for the
specification for all task conditions (although not all the correct
non-attention grabbing images are set). The task configuration also now
contains "label" states that it passes through for just one second to
show what state the system is in. Habituation aborting shows a
notification now.
-The correct stimuli are now uploaded (old stimuli placed on old_test
folder). Some stimuli images were resized for efficiency and the label
images were added.
-The MainFormConfiguration.json file was updated to include the
appropriate Block index values for the condition configurations. It
should now appropriate map through the whole experiment.