Release v1.1
-Added Second Coder field to input and logs
-Doubled the size of notification text (should be easier to read now)
-Moved a few scripts into their appropriate folders (no functional
changes)
-Added experimental "displaySwitch" command line argument which would
move the display from monitor 0 to monitor 1 between the form
submissions. Note, this changes the resolution to that monitor and the
resolution stays on restart if it is smaller than the main resolution.
The default monitor also sticks on restart. It is advised that the
application either be run in switch mode or not to avoid confusion, and
both monitors should be the same resolution (1920x1080).